Can generate frequency tables!
Added the runner because I'm getting excited to be done but also bummed that i'm still not done yet.

Oh also NoteConverter has NoteToNumber now for the frequency analysis, and the TuneEvaluator is damn close to being done frequency evaluation. Maybe this might work. Or maybe it'll all fall apart and I'll cry.